get matrix/macro from parser
