+++ .gitignore
@@ -20,3 +20,5 @@ data.json
 
 # Exclude macOS Finder (System Explorer) View States
 .DS_Store
+src/parser/latex-parser.terms.ts
+src/parser/latex-parser.ts
