+++ package.json
@@ -15,7 +15,8 @@
 		"version": "npm run lint && npm run build && node version-bump.mjs && git add manifest.json versions.json",
 		"lint": "npm run tsc && npm run eslint",
 		"tsc": "tsc --noEmit -skipLibCheck",
-		"eslint": "eslint --fix"
+		"eslint": "eslint --fix",
+		"generator": "lezer-generator src/parser/latex.grammar -o src/parser/latex-parser.ts --typeScript"
 	},
 	"keywords": [],
 	"author": "",
@@ -32,6 +33,7 @@
 		"@eslint/eslintrc": "^3.3.3",
 		"@eslint/js": "^9.39.2",
 		"@lezer/common": "^1.0.3",
+		"@lezer/generator": "^1.8.0",
 		"@lezer/highlight": "^1.2.3",
 		"@types/node": "^16.11.6",
 		"@typescript-eslint/eslint-plugin": "^8.55.0",
@@ -48,7 +50,8 @@
 		"typescript": "^5.9.3"
 	},
 	"dependencies": {
-		"valibot": "^1.3.1",
-		"@popperjs/core": "2.11.5"
+		"@lezer/lr": "^1.4.10",
+		"@popperjs/core": "2.11.5",
+		"valibot": "^1.3.1"
 	}
 }
