--- src/parser/tokens.mts
@@ -1,484 +1,0 @@
-/* Hand-written tokenizer for LaTeX. */
-
-import { ExternalTokenizer, ContextTracker } from '@lezer/lr'
-
-import {
-  LiteralArgContent,
-  SpaceDelimitedLiteralArgContent,
-  VerbContent,
-  VerbatimContent,
-  LstInlineContent,
-  Begin,
-  End,
-  KnownEnvironment,
-  Csname,
-  TrailingWhitespaceOnly,
-  TrailingContent,
-  RefCtrlSeq,
-  RefStarrableCtrlSeq,
-  CiteCtrlSeq,
-  CiteStarrableCtrlSeq,
-  LabelCtrlSeq,
-  MathTextCtrlSeq,
-  HboxCtrlSeq,
-  HrefCtrlSeq,
-  VerbCtrlSeq,
-  DefCtrlSeq,
-  LetCtrlSeq,
-  LeftCtrlSeq,
-  RightCtrlSeq,
-  NewCommandCtrlSeq,
-  RenewCommandCtrlSeq,
-  NewEnvironmentCtrlSeq,
-  RenewEnvironmentCtrlSeq,
-  DocumentEnvName,
-  TabularEnvName,
-  EquationEnvName,
-  EquationArrayEnvName,
-  VerbatimEnvName,
-  OpenParenCtrlSym,
-  CloseParenCtrlSym,
-  OpenBracketCtrlSym,
-  CloseBracketCtrlSym,
-  LineBreakCtrlSym,
-  TextColorCtrlSeq,
-  ColorBoxCtrlSeq,
-  HLineCtrlSeq,
-  ParBoxCtrlSeq,
-  // Marker for end of argument lists
-  endOfArguments,
-  hasMoreArguments,
-  hasMoreArgumentsOrOptionals,
-  endOfArgumentsAndOptionals,
-  TextBoldCtrlSeq,
-  TextItalicCtrlSeq,
-  TextTeletypeCtrlSeq,
-  TextSansSerifCtrlSeq,
-  EmphasisCtrlSeq,
-  UnderlineCtrlSeq,
-} from './latex.terms.mts'
-
-const MAX_ARGUMENT_LOOKAHEAD = 100
-
-function nameChar(ch: number) {
-  // we accept A-Z a-z 0-9 * + @ in environment names
-  return (
-    (ch >= 65 && ch <= 90) ||
-    (ch >= 97 && ch <= 122) ||
-    (ch >= 48 && ch <= 57) ||
-    ch === 42 ||
-    ch === 43 ||
-    ch === 64
-  )
-}
-
-// match [a-zA-Z]
-function alphaChar(ch: number) {
-  return (ch >= 65 && ch <= 90) || (ch >= 97 && ch <= 122)
-}
-
-let cachedName = null
-let cachedInput = null
-let cachedPos = 0
-function envNameAfter(input, offset) {
-  const pos = input.pos + offset
-  if (cachedInput === input && cachedPos === pos) {
-    return cachedName
-  }
-  if (input.peek(offset) !== '{'.charCodeAt(0)) return
-  offset++
-  let name = ''
-  for (; ;) {
-    const next = input.peek(offset)
-    if (!nameChar(next)) break
-    name += String.fromCharCode(next)
-    offset++
-  }
-  cachedInput = input
-  cachedPos = pos
-  return (cachedName = name || null)
-}
-
-function ElementContext(name, parent) {
-  this.name = name
-  this.parent = parent
-  this.hash = parent ? parent.hash : 0
-  for (let i = 0; i < name.length; i++)
-    this.hash +=
-      (this.hash << 4) + name.charCodeAt(i) + (name.charCodeAt(i) << 8)
-}
-
-export const elementContext = new ContextTracker({
-  start: null,
-  shift(context, term, stack, input) {
-    return term === Begin
-      ? new ElementContext(envNameAfter(input, '\\begin'.length) || '', context)
-      : context
-  },
-  reduce(context, term) {
-    return term === KnownEnvironment && context ? context.parent : context
-  },
-  reuse(context, node, _stack, input) {
-    const type = node.type.id
-    return type === Begin
-      ? new ElementContext(envNameAfter(input, 0) || '', context)
-      : context
-  },
-  hash(context) {
-    return context ? context.hash : 0
-  },
-  strict: false,
-})
-
-// tokenizer for \verb|...| commands
-export const verbTokenizer = new ExternalTokenizer(
-  (input, stack) => {
-    if (input.next === '*'.charCodeAt(0)) input.advance()
-    const delimiter = input.next
-    if (delimiter === -1) return // hit end of file
-    if (/\s|\*/.test(String.fromCharCode(delimiter))) return // invalid delimiter
-    input.advance()
-    for (; ;) {
-      const next = input.next
-      if (next === -1 || next === CHAR_NEWLINE) return
-      input.advance()
-      if (next === delimiter) break
-    }
-    return input.acceptToken(VerbContent)
-  },
-  { contextual: false }
-)
-
-// tokenizer for \lstinline|...| commands
-export const lstinlineTokenizer = new ExternalTokenizer(
-  (input, stack) => {
-    let delimiter = input.next
-    if (delimiter === -1) return // hit end of file
-    if (/\s/.test(String.fromCharCode(delimiter))) {
-      return // invalid delimiter
-    }
-    if (delimiter === CHAR_OPEN_BRACE) {
-      delimiter = CHAR_CLOSE_BRACE
-    }
-    input.advance()
-    for (; ;) {
-      const next = input.next
-      if (next === -1 || next === CHAR_NEWLINE) return
-      input.advance()
-      if (next === delimiter) break
-    }
-    return input.acceptToken(LstInlineContent)
-  },
-  { contextual: false }
-)
-
-const matchForward = (input, expected, offset = 0) => {
-  for (let i = 0; i < expected.length; i++) {
-    if (String.fromCharCode(input.peek(offset + i)) !== expected[i]) {
-      return false
-    }
-  }
-  return true
-}
-
-// tokenizer for \begin{verbatim}...\end{verbatim} environments
-export const verbatimTokenizer = new ExternalTokenizer(
-  (input, stack) => {
-    const delimiter = '\\end{' + stack.context.name + '}'
-    for (let offset = 0; ; offset++) {
-      const next = input.peek(offset)
-      if (next === -1 || matchForward(input, delimiter, offset)) {
-        return input.acceptToken(VerbatimContent, offset)
-      }
-    }
-  },
-  { contextual: false }
-)
-
-// tokenizer for \href{...} and similar commands
-export const literalArgTokenizer = new ExternalTokenizer(
-  input => {
-    for (let offset = 0; ; offset++) {
-      const next = input.peek(offset)
-      if (next === -1 || next === CHAR_CLOSE_BRACE) {
-        return input.acceptToken(LiteralArgContent, offset)
-      }
-    }
-  },
-  { contextual: false }
-)
-
-// tokenizer for literal content delimited by whitespace, such as in `\input foo.tex`
-export const spaceDelimitedLiteralArgTokenizer = new ExternalTokenizer(
-  input => {
-    for (let offset = 0; ; offset++) {
-      const next = input.peek(offset)
-      if (next === -1 || next === CHAR_SPACE || next === CHAR_NEWLINE) {
-        return input.acceptToken(SpaceDelimitedLiteralArgContent, offset)
-      }
-    }
-  },
-  { contextual: false }
-)
-
-// helper function to look up charCodes
-function _char(s) {
-  return s.charCodeAt(0)
-}
-
-const CHAR_BACKSLASH = _char('\\')
-const CHAR_OPEN_BRACE = _char('{')
-const CHAR_OPEN_BRACKET = _char('[')
-const CHAR_CLOSE_BRACE = _char('}')
-const CHAR_TAB = _char('\t')
-const CHAR_SPACE = _char(' ')
-const CHAR_NEWLINE = _char('\n')
-
-const lookaheadTokenizer = getToken =>
-  new ExternalTokenizer(
-    input => {
-      for (let i = 0; i < MAX_ARGUMENT_LOOKAHEAD; ++i) {
-        const next = input.peek(i)
-        if (next === CHAR_SPACE || next === CHAR_TAB) {
-          continue
-        }
-        const token = getToken(next)
-        if (token) {
-          input.acceptToken(token)
-          return
-        }
-      }
-    },
-    { contextual: false, fallback: true }
-  )
-
-export const argumentListTokenizer = lookaheadTokenizer(next => {
-  if (next === CHAR_OPEN_BRACE) {
-    return hasMoreArguments
-  } else {
-    return endOfArguments
-  }
-})
-
-export const argumentListWithOptionalTokenizer = lookaheadTokenizer(next => {
-  if (next === CHAR_OPEN_BRACE || next === CHAR_OPEN_BRACKET) {
-    return hasMoreArgumentsOrOptionals
-  } else {
-    return endOfArgumentsAndOptionals
-  }
-})
-
-const CHAR_AT_SYMBOL = _char('@')
-
-export const csnameTokenizer = new ExternalTokenizer((input, stack) => {
-  let offset = 0
-  let end = -1
-  // look at the first character, we are looking for acceptable control sequence names
-  // including @ signs, \\[a-zA-Z@]+
-  const next = input.peek(offset)
-  if (next === -1) {
-    return
-  }
-  // reject anything not starting with a backslash,
-  // we only accept control sequences
-  if (next !== CHAR_BACKSLASH) {
-    return
-  }
-  offset++
-  for (; ;) {
-    const next = input.peek(offset)
-    // stop when we reach the end of file or a non-csname character
-    if (next === -1 || !(alphaChar(next) || next === CHAR_AT_SYMBOL)) {
-      end = offset - 1
-      break
-    }
-    end = offset
-    offset++
-  }
-  if (end === -1) return
-  // accept the content as a valid control sequence
-  return input.acceptToken(Csname, end + 1)
-})
-
-const END_DOCUMENT_MARK = '\\end{document}'.split('').reverse()
-
-export const trailingContentTokenizer = new ExternalTokenizer(
-  (input, stack) => {
-    if (input.next === -1) return // no trailing content
-    // Look back for end-document mark, bail out if any characters do not match
-    for (let i = 1; i < END_DOCUMENT_MARK.length + 1; i++) {
-      if (String.fromCharCode(input.peek(-i)) !== END_DOCUMENT_MARK[i - 1]) {
-        return
-      }
-    }
-    while (input.next === CHAR_SPACE || input.next === CHAR_NEWLINE) {
-      const next = input.advance()
-      if (next === -1) return input.acceptToken(TrailingWhitespaceOnly) // trailing whitespace only
-    }
-    // accept the all content up to the end of the document
-    while (input.advance() !== -1) {
-      //
-    }
-    return input.acceptToken(TrailingContent)
-  }
-)
-
-const refCommands = new Set([
-  '\\eqref',
-])
-
-const refStarrableCommands = new Set([
-  '\\ref',
-])
-
-const citeCommands = new Set([
-])
-
-const citeStarredCommands = new Set([
-])
-
-
-
-
-
-const labelCommands = new Set([
-	"\\label",
-]);
-
-const mathTextCommands = new Set([
-	"\\text",
-	"\\tag",
-	"\\textrm",
-]);
-
-const otherKnowncommands = {
-  '\\hbox': HboxCtrlSeq,
-  '\\href': HrefCtrlSeq,
-  '\\verb': VerbCtrlSeq,
-  '\\def': DefCtrlSeq,
-  '\\let': LetCtrlSeq,
-  '\\left': LeftCtrlSeq,
-  '\\right': RightCtrlSeq,
-  '\\newcommand': NewCommandCtrlSeq,
-  '\\renewcommand': RenewCommandCtrlSeq,
-  '\\newenvironment': NewEnvironmentCtrlSeq,
-  '\\renewenvironment': RenewEnvironmentCtrlSeq,
-  '\\textcolor': TextColorCtrlSeq,
-  '\\colorbox': ColorBoxCtrlSeq,
-  '\\hline': HLineCtrlSeq,
-  '\\parbox': ParBoxCtrlSeq,
-  '\\textbf': TextBoldCtrlSeq,
-  '\\textit': TextItalicCtrlSeq,
-  '\\texttt': TextTeletypeCtrlSeq,
-  '\\textsf': TextSansSerifCtrlSeq,
-  '\\emph': EmphasisCtrlSeq,
-  '\\underline': UnderlineCtrlSeq,
-}
-// specializer for control sequences
-// return new tokens for specific control sequences
-export const specializeCtrlSeq = (name: string, terms: string) => {
-  if (name === '\\begin') return Begin
-  if (name === '\\end') return End
-  if (refCommands.has(name)) {
-    return RefCtrlSeq
-  }
-  if (refStarrableCommands.has(name)) {
-    return RefStarrableCtrlSeq
-  }
-  if (citeCommands.has(name)) {
-    return CiteCtrlSeq
-  }
-  if (citeStarredCommands.has(name)) {
-    return CiteStarrableCtrlSeq
-  }
-  if (labelCommands.has(name)) {
-    return LabelCtrlSeq
-  }
-  if (mathTextCommands.has(name)) {
-    return MathTextCtrlSeq
-  }
-  return otherKnowncommands[name] || -1
-}
-
-const tabularEnvNames = new Set([
-])
-
-const equationEnvNames = new Set([
-  'equation',
-  'equation*',
-  'displaymath',
-  'math',
-  'multline',
-  'multline*',
-  'matrix',
-])
-
-const equationArrayEnvNames = new Set([
-  'array',
-  'eqnarray',
-  'eqnarray*',
-  'align',
-  'align*',
-  'alignat',
-  'alignat*',
-  'flalign',
-  'flalign*',
-  'gather',
-  'gather*',
-  'pmatrix',
-  'pmatrix*',
-  'bmatrix',
-  'bmatrix*',
-  'Bmatrix',
-  'Bmatrix*',
-  'vmatrix',
-  'vmatrix*',
-  'Vmatrix',
-  'Vmatrix*',
-  'smallmatrix',
-  'smallmatrix*',
-  'split',
-  'gathered',
-  'aligned',
-  'alignedat',
-  'cases',
-  'cases*',
-  'dcases',
-  'dcases*',
-  'rcases',
-  'rcases*',
-])
-
-const verbatimEnvNames = new Set([
-])
-
-const otherKnownEnvNames = {
-}
-
-export const specializeEnvName = (name: string, terms: string) => {
-  if (tabularEnvNames.has(name)) {
-    return TabularEnvName
-  }
-  if (equationEnvNames.has(name)) {
-    return EquationEnvName
-  }
-  if (equationArrayEnvNames.has(name)) {
-    return EquationArrayEnvName
-  }
-  if (verbatimEnvNames.has(name)) {
-    return VerbatimEnvName
-  }
-  return otherKnownEnvNames[name] || -1
-}
-
-const otherKnownCtrlSyms = {
-  '\\(': OpenParenCtrlSym,
-  '\\)': CloseParenCtrlSym,
-  '\\[': OpenBracketCtrlSym,
-  '\\]': CloseBracketCtrlSym,
-  '\\\\': LineBreakCtrlSym,
-}
-
-export const specializeCtrlSym = (name: string, terms: string) => {
-  return otherKnownCtrlSyms[name] || -1
-}
